Fix docker context build
The `.github/docker` and debian directories are copied into the Docker
context from git in order to ensure they're pristine.  If the debian
directory has been created by a script but doesn't exist in git, it
breaks the script.

Fix this by checking if the directory exists in git, not if it exists
on the filesystem, and create manually in Docker context so the
`Dockerfile` `COPY` command succeeds.

Also, print Docker context directories as well as files to help debug
situations like this.