Get cloudsmith namespace from environment or config file
The old scheme of getting the Cloudsmith namespace from the username
didn't allow for org namespaces, and now that Cloudsmith ONLY allows
org namespaces, the old scheme completely broke.

This patch takes the guessing out by requiring the Cloudsmith
namespace to be explicitly set in `$CLOUDSMITH_NAMESPACE` (probably in
the GH repo settings) or in `.github/debian-distro-settings.yaml`.